+++ Models/ViewModels/Orders/BillViewModel.cs
@@ -35,7 +35,7 @@ namespace QuanLyBanHangCore.Models.ViewModels
     public class ItemOfBillViewModel
     {
         public ulong Gia { get; set; }
-        public ushort SoLuong { get; set; }
+        public uint SoLuong { get; set; }
         public string Ten { get; set; }
         public ulong LayTamTinh()
         {
